requirements.txt
boto3>=1.28.0
pandas>=2.0.0
requests>=2.31.0
python-dotenv>=1.0.0
tqdm>=4.66.0
